requirements.txt
annotated-types==0.6.0
ascii-magic==2.3.0
black==23.12.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
framegrab==0.4.3
frozendict==2.4.0
groundlight==0.13.2
idna==3.6
imgcat==0.5.0
mypy-extensions==1.0.0
numpy==1.26.3
opencv-python==4.9.0.80
packaging==23.2
pathspec==0.12.1
Pillow==9.5.0
platformdirs==4.1.0
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
typer==0.9.0
typing_extensions==4.9.0
urllib3==1.26.18